third_party/blink/web_tests/fast/dom/MutationObserver/disconnect-cancel-pending-expected.txt

Test that WebKitMutationObserver.disconnect cancels pending delivery

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Disconnecting should cancel any pending delivery...
PASS mutations is null
...and re-observing should not see any of the previously-generated records.
PASS mutations.length is 1
PASS mutations[0].attributeName is "bar"
PASS successfullyParsed is true

TEST COMPLETE
